[CLOUDTRUST-2439] Allow to run static analysis on child modules.
For that to work, without editing the cloudtrust-parent pom too much,
weh have to recurse in each child pom and run
spotbugs/pmd/dependency-check.
This brings possible issues with a shared volume of maven dependencies
in the CI (race concurrency, cleanup).

Also, do only call deploy when we build master.